requirements.txt
asgiref==3.9.1
Django==5.2.4
mysqlclient==2.2.7
python-decouple==3.8
sqlparse==0.5.3
gunicorn==21.2.0